test: refactor richtext-resolver tests
